requirements.txt
appdirs==1.4.0
Django==1.10.5
django-datetime-widget==0.9.3
django-widget-tweaks==1.4.1
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
six==1.10.0
